Introduce syscall for managing xattr exception list
Introduce {add/remove/list}xattrexception system calls that allow
to manage extended attribute write exceptions for a user
namespace. This is useful in case CAP_SYS_ADMIN is not set, yet
the user may need to set certain extended attribute. The lack
of CAP_SYS_ADMIN in a namespace would normally prevent the setting of
all extended attributes.

The system calls must be executed by a parent user namespace and provide
the file descriptor of a child user namespace that is to be granted the
exception. The granting user namespace must have CAP_SYS_ADMIN or must
have the exception itself for it to be able to grant it to another
user namespace. This prevents gaining such privileges in a newly spawned
user namespace. There is no inheritance of these exceptions. Only the
root user can execute this syscall.

Currently only security.ima is allowed as an exception.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>