xattr: Enable security.capability in user namespaces
This patch enables security.capability in user namespaces but also
takes a more general approach to enabling extended attributes in user
namespaces.

The following rules describe the approach using security.foo as a
'user namespace enabled' extended attribute:

Reading of extended attributes:

1) Reading security.foo from a user namespace will read
   security.foo@uid=<uid> instead with uid being the mapping of root.
   An exception is if root is mapped to uid 0 on the host, and in this
   case we will read security.foo directly.
   All security.foo@uid=<uid> can be read from the host with values
   of <uid> being subject to range checking.
2) All security.foo@uid=<uid> with valid uid mapping in the user
   namespace can be read.
3) No other security.foo* can be read.

The same rules for reading apply to writing and removing of user
namespace enabled extended attributes.

When listing extended attributes of a file, only those are presented
to the user namespace that have a valid mapping. Besides that, names
of the extended attributes are adjusted to represent the mapping.
This means that if root is mapped to uid 1000 on the host, the
security.foo@uid=1000 will be listed as security.foo in the user
namespace, security.foo@uid=1001 becomes security.foo@uid=1 and so on.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>