tpm: ibmvtpm: Avoid -EINTR error when IMA talks to TPM
When IMA is taking measurements during compilation for example and a
user presses ctrl-c to abort the compilation, lots of these types of
messages will appear in the kernel log:

[ 7406.275163] tpm tpm0: tpm_transmit: tpm_recv: error -4
[ 7406.275242] ima: Error Communicating to TPM chip, result: -4

The issue is caused by the fact that the IBM vTPM driver's recv()
function is called immediately after send() without waiting for
status on whether a response was received. It currently waits for
the current command to finish using this call that ends up throwing
these error messages because it is 'interruptible':

sig = wait_event_interruptible(ibmvtpm->wq,
                               !ibmvtpm->tpm_processing_cmd);

Instead, it should be using the polling loop in tpm_try_transmit()
that uses a command's duration to poll until a result has been
returned by the TPM, thus ending when the timeout has occurred but
not responding to users' ctrl-c request anymore. To stay in this
polling loop we now extend tpm_ibmvtpm_status() to return
PM_STATUS_BUSY for as long as the vTPM is busy. Since we will need
the timeouts in this loop now we get the TPM 1.2 and TPM 2 timeouts
with tpm_get_timeouts().

We change tpm_processing_cmd to tpm_status and set the TPM_STATUS_BUSY
flag while the vTPM is busy processing a command.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>
Fixes: 18b3670d79ae9 ("tpm: ibmvtpm: Add support for TPM2")
Cc: Nayna Jain <nayna@linux.ibm.com>
Cc: George Wilson <gcwilson@linux.ibm.com>