tpm/kexec: Protect the measurement log of the of-tree
The memory area of the TPM measurement log is currently not properly
protected across kexec when an Open Firmware Devicetree is used and
therefore the contents of the log get corrupted. Fix this for the
kexec_file_load() syscall by allocating a buffer and copying the
contents of the existing log into it. The new buffer is preserved
across the kexec and a pointer to it is available when the new kernel is
started. To achieve this, store the allocated buffer's address in the
flattened device tree (fdt) under the name linux,tpm-kexec-buffer and
search for this entry early in the kernel startup before the TPM subsystem
starts up. Adjust the pointer in the of-tree stored under linux,sml-base
to point to this buffer holding the preserved log. The TPM driver can then
read the base address from this entry when making the log available.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>